+++ mypy.ini
@@ -2220,6 +2220,14 @@ disallow_untyped_defs = false
 disallow_incomplete_defs = false
 check_untyped_defs = false
 
+[mypy-ddtrace.contrib.internal.django.routing]
+ignore_errors = true
+disallow_untyped_calls = false
+disallow_untyped_defs = false
+disallow_incomplete_defs = false
+check_untyped_defs = false
+warn_return_any = false
+
 [mypy-ddtrace.contrib.internal.django.restframework]
 disallow_untyped_defs = false
 disallow_incomplete_defs = false
